# services/api/openapi.yaml
openapi: 3.0.3
info:
  title: tyr.bet API
  version: 1.0.0
  description: "Testnet-only. Session auth = signed `tyr_session` cookie from the passkey routes. Agent routes are signed with the agent key (x-tyr-agent / x-tyr-timestamp / x-tyr-signature, EIP-191 over method, path, timestamp, sha256(body)); paid agent routes answer 402 with an MPP challenge. Live events: GET /ws (see WsEvent)."
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: tyr_session
    agent:
      type: apiKey
      in: header
      name: x-tyr-signature
    admin:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
          description: machine-readable reason (e.g. limit, balance, region)
        issues:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
            required:
              - path
              - message
            additionalProperties: false
          description: request validation failures
      required:
        - error
      additionalProperties: false
      description: Error
...
    ReceiptSummary:
      type: object
      properties:
        id:
          type: string
        kind:
          anyOf:
            - type: string
              enum:
                - bet
                - zcash-payout
                - bet+hedge
            - type: string
        subjectId:
          type: string
          nullable: true
        payloadHash:
          type: string
        visibility:
          type: string
          enum:
            - private
            - public
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - kind
        - subjectId
        - payloadHash
        - visibility
        - createdAt
      additionalProperties: false
    Verification:
      type: object
      properties:
        receiptId:
          type: string
        kind:
          type: string
        ok:
          type: boolean
        checks:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              ok:
                type: boolean
              detail:
                type: string
            required:
              - name
              - ok
            additionalProperties: false
      required:
        - receiptId
        - kind
        - ok
        - checks
      additionalProperties: false
    Proof:
      type: object
      properties:
        receiptId:
          type: string
        kind:
          type: string
        payloadHash:
          type: string
        provedAt:
          type: string
          format: date-time
        tempo:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              tx:
                type: string
              memo:
                type: string
              explorer:
                type: string
            required:
              - label
              - tx
              - memo
              - explorer
            additionalProperties: false
        solana:
          type: object
          properties:
            settleTx:
              type: string
              nullable: true
            opening:
              type: object
              properties:
                amountUnits:
                  type: string
                salt:
                  type: string
              required:
                - amountUnits
                - salt
              additionalProperties: false
              nullable: true
            attestations:
              type: array
              items:
                type: object
                properties:
                  label:
                    type: string
                  tx:
                    type: string
                  explorer:
                    type: string
                  attestation:
                    type: object
                    properties:
                      statement:
                        type: object
                        properties: {}
                        additionalProperties: true
                    required:
                      - statement
                    additionalProperties: true
                required:
                  - label
                  - tx
                  - explorer
                  - attestation
                additionalProperties: false
          required:
            - settleTx
            - opening
            - attestations
          additionalProperties: false
          nullable: true
        zcash:
          type: object
          properties:
            kind:
              type: string
              enum:
                - zcash-ovk-view+frost
            receiptId:
              type: string
            orderTxid:
              type: string
            payoutTxid:
              type: string
            view:
              type: object
              properties: {}
              additionalProperties: true
              nullable: true
            instruction:
              type: object
              properties: {}
              additionalProperties: true
            frost:
              type: object
              properties:
                signature:
                  type: string
                signers:
                  type: string
                  nullable: true
                verifyingKey:
                  type: string
                threshold:
                  type: string
                valid:
                  type: boolean
              required:
                - signature
                - signers
                - verifyingKey
                - threshold
                - valid
              additionalProperties: false
            matches:
              type: boolean
          required:
            - kind
            - receiptId
            - orderTxid
            - payoutTxid
            - view
            - instruction
            - frost
            - matches
          additionalProperties: false
          nullable: true
      required:
        - receiptId
        - kind
        - payloadHash
        - provedAt
        - tempo
        - solana
        - zcash
      additionalProperties: false
...
paths:
...
  /api/zcash/request:
    post:
      tags:
        - zcash
      description: ZIP-321 payment request for a shielded bet (Flow B). No account needed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                outcome:
                  type: integer
                  minimum: 0
                side:
                  type: string
                  enum:
                    - "yes"
                    - "no"
                stakeUsd:
                  type: number
                  exclusiveMinimum: true
                  minimum: 0
                  maximum: 10000
                returnUA:
                  type: string
                  pattern: ^u(1|test1|regtest1)[0-9a-z]+$
                  maxLength: 255
              required:
                - outcome
                - side
                - stakeUsd
                - returnUA
              additionalProperties: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  uri:
                    type: string
                    description: ZIP-321 payment URI (render as QR)
                  to:
                    type: string
                  zat:
                    type: string
                  memo:
                    type: string
                  zecUsd:
                    type: number
                required:
                  - uri
                  - to
                  - zat
                  - memo
                  - zecUsd
                additionalProperties: false
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/zcash/orders/{txid}:
    get:
      tags:
        - zcash
      parameters:
        - schema:
            type: string
            minLength: 1
          in: path
          name: txid
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  txid:
                    type: string
                  status:
                    type: string
                  marketId:
                    type: string
                  side:
                    type: string
                    enum:
                      - "yes"
                      - "no"
                  size:
                    type: string
                    description: decimal number as a string (exact, from Postgres numeric)
                  zecUsd:
                    type: string
                    nullable: true
                    description: decimal number as a string (exact, from Postgres numeric)
                  payoutZat:
                    type: string
                    nullable: true
                  payoutTxid:
                    type: string
                    nullable: true
                  error:
                    type: string
                    nullable: true
                  orderId:
                    type: string
                    nullable: true
                required:
                  - id
                  - txid
                  - status
                  - marketId
                  - side
                  - size
                  - zecUsd
                  - payoutZat
                  - payoutTxid
                  - error
                  - orderId
                additionalProperties: false
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/zcash/orders/{txid}/disclosure:
    get:
      tags:
        - zcash
      description: OVK view of the payout + FROST 2-of-3 check (the txid is the capability).
      parameters:
        - schema:
            type: string
            minLength: 1
          in: path
          name: txid
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  kind:
                    type: string
                    enum:
                      - zcash-ovk-view+frost
                  receiptId:
                    type: string
                  orderTxid:
                    type: string
                  payoutTxid:
                    type: string
                  view:
                    type: object
                    properties: {}
                    additionalProperties: true
                    nullable: true
                  instruction:
                    type: object
                    properties: {}
                    additionalProperties: true
                  frost:
                    type: object
                    properties:
                      signature:
                        type: string
                      signers:
                        type: string
                        nullable: true
                      verifyingKey:
                        type: string
                      threshold:
                        type: string
                      valid:
                        type: boolean
                    required:
                      - signature
                      - signers
                      - verifyingKey
                      - threshold
                      - valid
                    additionalProperties: false
                  matches:
                    type: boolean
                required:
                  - kind
                  - receiptId
                  - orderTxid
                  - payoutTxid
                  - view
                  - instruction
                  - frost
                  - matches
                additionalProperties: false
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/zcash/orders/{txid}/receipt:
    get:
      tags:
        - zcash
      parameters:
        - schema:
            type: string
            minLength: 1
          in: path
          name: txid
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  payloadHash:
                    type: string
                  payload: {}
                  verification:
                    $ref: "#/components/schemas/Verification"
                required:
                  - id
                  - payloadHash
                  - verification
                additionalProperties: false
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/receipts:
    get:
      tags:
        - receipts
      security:
        - session: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  receipts:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReceiptSummary"
                required:
                  - receipts
                additionalProperties: false
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/receipts/{id}:
    get:
      tags:
        - receipts
      description: Receipt payload + hash. Owner, or anyone if the receipt is public; 404 otherwise.
      parameters:
        - schema:
            type: string
            minLength: 1
          in: path
          name: id
          required: true
      security:
        - {}
        - session: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  kind:
                    anyOf:
                      - type: string
                        enum:
                          - bet
                          - zcash-payout
                          - bet+hedge
                      - type: string
                  subjectId:
                    type: string
                    nullable: true
                  payloadHash:
                    type: string
                  visibility:
                    type: string
                    enum:
                      - private
                      - public
                  createdAt:
                    type: string
                    format: date-time
                  payload: {}
                required:
                  - id
                  - kind
                  - subjectId
                  - payloadHash
                  - visibility
                  - createdAt
                additionalProperties: false
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/receipts/{id}/proof:
    get:
      tags:
        - receipts
      description: "Prove bundle: Tempo tx + memo, Solana commitment opening + auditor attestation, Zcash OVK view + FROST. Owner, or anyone if the receipt is public; 404 otherwise."
      parameters:
        - schema:
            type: string
            minLength: 1
          in: path
          name: id
          required: true
      security:
        - {}
        - session: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proof"
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/receipts/{id}/verify:
    get:
      tags:
        - receipts
      description: Re-checks the receipt against the chains (not the DB). Owner, or anyone if the receipt is public; 404 otherwise.
      parameters:
        - schema:
            type: string
            minLength: 1
          in: path
          name: id
          required: true
      security:
        - {}
        - session: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Verification"
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/receipts/{id}/visibility:
    put:
      tags:
        - receipts
      description: Publish (proves first, so the public page serves the attested bundle) or hide.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                visibility:
                  type: string
                  enum:
                    - public
                    - private
              required:
                - visibility
              additionalProperties: false
      parameters:
        - schema:
            type: string
            minLength: 1
          in: path
          name: id
          required: true
      security:
        - session: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiptSummary"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"